pyproject.toml
[tool.poetry]
name = "capstone-be"
version = "0.1.0"
description = ""
authors = ["chaechaeis <hch011007@ewhain.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.2.2"
django-extensions = "^3.2.3"
pillow = "^10.4.0"
gunicorn = "^23.0.0"
uvicorn = "^0.30.6"
boto3 = "^1.35.19"
django-cors-headers = "^4.4.0"
python-dotenv = "^1.0.1"
django-storages = {extras = ["s3"], version = "^1.14.4"}
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.8.0"
mypy = "^1.11.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88
known_third_party = ["example_package"] # 필요에 따라 third-party 패키지를 여기에 추가
skip_gitignore = true
